adk_agents/game_round/requirements.txt
google-adk>=0.1.0
google-cloud-aiplatform
google-auth
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0 